+++ android/src/main/kotlin/lab/neruno/guardian_liveness/activity/LivenessActivity.java
@@ -14,7 +14,6 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
 
-import ai.advance.common.utils.LogUtil;
 import ai.advance.common.utils.ScreenUtil;
 import ai.advance.liveness.lib.GuardianLivenessDetectionSDK;
 import ai.advance.liveness.lib.LivenessResult;