+++ example/test/widget_test.dart
@@ -13,7 +13,7 @@ import 'package:guardian_liveness_example/main.dart';
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(MyApp());
+    await tester.pumpWidget(DemoLivenessApp());
 
     // Verify that platform version is retrieved.
     expect(