+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.1+2
+
+* Re-checked result type from "detectLiveness" method
+* Fixed Gradle dependencies for the Android Project
+* Updated Example App
+
 ## 0.0.1
 
 * Initial Commit. Built for Android only.